fix: update build.started env

+++ plugin.go
@@ -41,8 +41,8 @@ type (
 		Message  string
 		Status   string
 		Link     string
-		Started  float64
-		Finished float64
+		Started  int64
+		Finished int64
 		DeployTo string
 	}
